oops floats

+++ game/opengl_text.hh
@@ -14,9 +14,9 @@ struct TZoomText {
 	/// text
 	std::string string;
 	/// zoomfactor
-	double factor;
+	float factor;
 	/// constructor
-	TZoomText(std::string const& str = std::string()): string(str), factor(1.0) {}
+	TZoomText(std::string const& str = std::string()): string(str), factor(1.0f) {}
 };
 
 /// Special theme for creating opengl themed surfaces
@@ -38,9 +38,9 @@ struct TextStyle {
 	};
 	Color fill_col; ///< fill color
 	Color stroke_col; ///< stroke color
-	double stroke_width; ///< stroke thickness
-	double stroke_miterlimit; ///< stroke miter limit
-	double fontsize; ///< fontsize
+	float stroke_width; ///< stroke thickness
+	float stroke_miterlimit; ///< stroke miter limit
+	float fontsize; ///< fontsize
 	std::string fontfamily; ///< fontfamily
 	std::string fontstyle; ///< fontstyle
 	std::string fontweight; ///< fontweight
@@ -48,7 +48,7 @@ struct TextStyle {
 	std::string	stroke_linejoin; ///< stroke line-join type
 	std::string	stroke_linecap; ///< stroke line-join type
 	std::string text; ///< text
-	TextStyle(): stroke_width(), stroke_miterlimit(1.0), fontsize() {}
+	TextStyle(): stroke_width(), stroke_miterlimit(1.0f), fontsize() {}
 };
 	
 /// Convenience container for deciding how a given OpenGLText instance will be wrapped, ellipsized or fitted to the display area.
@@ -100,9 +100,9 @@ public:
 	/// draws full texture
 	void draw();
 	/// @return x
-	double x() const { return m_x; }
+	float x() const { return m_x; }
 	/// @return y
-	double y() const { return m_y; }
+	float y() const { return m_y; }
 	/// @returns dimension of texture
 	Dimensions& dimensions() { return m_texture.dimensions; }
 	/// @return number of lines rendered.
@@ -110,8 +110,8 @@ public:
 
 private:
 	size_t m_lines = 1;
-	double m_x;
-	double m_y;
+	float m_x;
+	float m_y;
 	Texture m_texture;
 };
 
@@ -119,7 +119,7 @@ private:
 class SvgTxtThemeSimple {
 public:
 	/// constructor
-	SvgTxtThemeSimple(fs::path const& themeFile, double factor = 1.0, WrappingStyle _wrapping = WrappingStyle().menuScreenText());
+	SvgTxtThemeSimple(fs::path const& themeFile, float factor = 1.0f, WrappingStyle _wrapping = WrappingStyle().menuScreenText());
 	/// renders text
 	void render(std::string _text);
 	/// draws texture
@@ -133,7 +133,7 @@ private:
 	std::unique_ptr<OpenGLText> m_opengl_text;
 	std::string m_cache_text;
 	TextStyle m_text;
-	double m_factor;
+	float m_factor;
 	WrappingStyle m_wrapping;
 };
 
@@ -187,7 +187,7 @@ public:
 	/// dimensions, what else
 	Dimensions dimensions;
 	/// constructor
-	SvgTxtTheme(fs::path const& themeFile, double factor = 1.0, WrappingStyle _wrapping = WrappingStyle().menuScreenText());
+	SvgTxtTheme(fs::path const& themeFile, float factor = 1.0f, WrappingStyle _wrapping = WrappingStyle().menuScreenText());
 	/// draws text with alpha
 	void draw(std::vector<TZoomText>& _text, bool padSyllables = false);
 	/// draw text with alpha
@@ -195,9 +195,9 @@ public:
 	/// sets highlight
 	void setHighlight(fs::path const& themeFile);
 	/// width
-	double w() const { return m_texture_width; }
+	float w() const { return m_texture_width; }
 	/// height
-	double h() const { return m_texture_height; }
+	float h() const { return m_texture_height; }
 	/// set align
 	void setAlign(Align align) { m_align = align; }
 	/// Returns the maximum number of lines in a contained OpenGLText.
@@ -206,13 +206,13 @@ public:
 private:
 	std::vector<std::unique_ptr<OpenGLText>> m_opengl_text;
 	Align m_align;
-	double m_x;
-	double m_y;
-	double m_width;
-	double m_height;
-	double m_factor;
-	double m_texture_width;
-	double m_texture_height;
+	float m_x;
+	float m_y;
+	float m_width;
+	float m_height;
+	float m_factor;
+	float m_texture_width;
+	float m_texture_height;
 	std::string m_cache_text;
 	TextStyle m_text;
 	TextStyle m_text_highlight;
